old/requirements.txt
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
Click==7.0
cloudpickle==0.8.0
cycler==0.10.0
dask==1.1.3
-e git://github.com/h1-the-swan/db_connect_mag.git@c1233fe1419291ec0d18dbd5d9f93d9675270f60#egg=db_connect_mag
decorator==4.2.1
distributed==1.26.0
-e git://github.com/h1-the-swan/h1theswan_utils.git@7049aaa30f8864e9eb23d94b50cc9b95e1740c8c#egg=h1theswan_utils
HeapDict==1.0.0
humanfriendly==4.10
idna==2.6
ipykernel==4.8.2
ipython==6.3.1
ipython-genutils==0.2.0
jedi==0.11.1
jupyter-client==5.2.3
jupyter-core==4.4.0
kiwisolver==1.0.1
locket==0.2.0
matplotlib==2.2.2
msgpack==0.6.1
-e git://github.com/h1-the-swan/mysql_connect.git@cb225c6d4f1c14bcd9275f1e934d6a007d576f97#egg=mysql_connect
numpy==1.16.1
pandas==0.24.1
parso==0.1.1
partd==0.3.9
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psutil==5.6.0
ptyprocess==0.5.2
py4j==0.10.7
pyarrow==0.13.0
Pygments==2.2.0
PyMySQL==0.8.0
pyparsing==2.2.0
pyspark==2.4.2
python-dateutil==2.7.5
python-dotenv==0.8.2
pytz==2018.9
PyYAML==3.13
pyzmq==17.0.0
requests==2.18.4
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
simplegeneric==0.8.1
six==1.12.0
sklearn==0.0
sortedcontainers==2.1.0
SQLAlchemy==1.2.7
tblib==1.3.2
toolz==0.9.0
tornado==5.0.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
zict==0.1.3
